fix: Resolve concurrency bug during dereferencing
By reusing the same refParser instance for all calls of the parse-function, it can come to concurrency issues if two parse-calls are executed at the same time (i.e. inside a Promise.all() call). This PR resolves the issue by creating a new json-schema-ref-parser for each call of the parse function.